+++ server/src/IoNameDatFile.ts
@@ -4,6 +4,7 @@ import {
 	HoverParams,
 	FoldingRangeParams,
 	FoldingRange,
+	integer,
 } from 'vscode-languageserver/node';
 
 import {
@@ -19,6 +20,8 @@ import { RobotControllerFile } from './RobotControllerFile';
 import * as Util from './Util';
 
 export class IoNameDatFile extends RobotControllerFile {
+
+	ioNameTable: Map<number, string> | undefined;
 
 	updateSection() {
 		if( this.sectionedDocument ) {
@@ -69,6 +72,74 @@ export class IoNameDatFile extends RobotControllerFile {
 		return this.sectionedDocument;
 	}
 
+	
+	updateIoName() {
+		if( this.ioNameTable ) {
+			return;
+		}
+		this.ioNameTable = new Map<number, string>();
+
+		const sectionedDocument = this.updateSection();
+		if( !sectionedDocument ) {
+			return null;
+		}
+
+		const textLines = this.workspace.getTextLines( this.filePath );
+
+		if( !textLines ) {
+			return;
+		}
+
+		const typeNumberTable = new Map<string, number>([
+			["IN", 0],
+			["OUT", 10000],
+			["EXIN", 20000],
+			["EXOUT", 30000],
+		]);
+
+		const sectionNames = sectionedDocument.getSectionNames();
+		sectionNames.forEach( ( sectionName ) => {
+			const section = sectionedDocument.getSection( sectionName );
+
+			if( !section ) {
+				return;
+			}
+
+			const typeNumber = typeNumberTable.get(sectionName);
+
+			if(typeNumber == undefined) {
+				// error: unknown io type
+				return;
+			}
+
+
+			for( let lineNo = section.contents.start.line; lineNo < section.contents.end.line; lineNo++ ) {
+				const text = textLines[lineNo];
+
+				const names = text.split(",");
+
+				const offset = lineNo - section.contents.start.line;
+
+				names.forEach( (name, i ) => {
+					if( name.length > 0 ) {
+						const index = offset * 4 + i;
+						const groupNumber = Math.floor( index / 8 ) + 1;
+						const bitNumber = index % 8;
+						const logicNumber = typeNumber + groupNumber * 10 + bitNumber; 
+						this.ioNameTable?.set( logicNumber, name );	
+					}
+				} );
+
+			}
+			
+		});
+	}
+
+	getIoName( logicalNumber: number ): string | undefined {
+		this.updateIoName();
+
+		return this.ioNameTable?.get( logicalNumber );
+	}
 
 	onHover(hoverParams: HoverParams): Hover | null {
 		const sectionedDocument = this.updateSection();